отрефачил структуру модуля с тестами
