+++ selenide-tests/src/test/java/com/tests/data/Blocker.java
@@ -1,4 +1,4 @@
-package com.tests;
+package com.tests.data;
 
 import org.junit.jupiter.api.Tag;
