+++ selenide-tests/src/test/java/com/tests/lesson3_selenideTests/GithubWebTests.java
@@ -1,7 +1,7 @@
-package com.github;
+package com.tests.lesson3_selenideTests;
 
 import com.codeborne.selenide.selector.ByAttribute;
-import com.demoQA.practiceForm.utils.TestBaseDemoQA;
+import com.tests.lesson9_PageObject.utils.TestBaseDemoQA;
 import io.qameta.allure.Description;
 import org.junit.jupiter.api.Test;
 
@@ -9,7 +9,8 @@ import static com.codeborne.selenide.Condition.text;
 import static com.codeborne.selenide.Condition.visible;
 import static com.codeborne.selenide.Selectors.byTagAndText;
 import static com.codeborne.selenide.Selectors.byText;
-import static com.codeborne.selenide.Selenide.*;
+import static com.codeborne.selenide.Selenide.$;
+import static com.codeborne.selenide.Selenide.open;
 
 
 public class GithubWebTests extends TestBaseDemoQA {
