+++ selenide-tests/src/test/java/com/tests/lesson9_PageObject/pages/components/CalendarComponent.java
@@ -1,4 +1,4 @@
-package com.demoQA.practiceForm.pages.components;
+package com.tests.lesson9_PageObject.pages.components;
 
 import static com.codeborne.selenide.Selectors.byText;
 import static com.codeborne.selenide.Selenide.$;
