+++ selenide-tests/src/test/java/com/tests/lesson9_PageObject/tests/TextBoxTests.java
@@ -1,16 +1,16 @@
-package com.demoQA.practiceForm.tests;
+package com.tests.lesson9_PageObject.tests;
 
+import com.tests.lesson9_PageObject.pages.TextBoxFormPage;
+import com.tests.lesson9_PageObject.utils.TestBaseDemoQA;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import com.demoQA.practiceForm.utils.TestBaseDemoQA;
 
 import static com.codeborne.selenide.Condition.text;
 import static com.codeborne.selenide.Selenide.$;
-import static com.demoQA.practiceForm.pages.TextBoxFormPage.*;
 
 public class TextBoxTests extends TestBaseDemoQA {
 
-    FormData formData = new FormData(
+    TextBoxFormPage.FormData formData = new TextBoxFormPage.FormData(
             "Yaroslav Radzievskiy",
             "Auto@test.qa",
             "Moscow, Mira st. 4",
@@ -20,8 +20,8 @@ public class TextBoxTests extends TestBaseDemoQA {
     @DisplayName("Проверка заполнения формы")
     void textBoxFormTest() {
         openPage("/text-box");
-        fillTestBoxForm(formData);
-        submitForm();
+        TextBoxFormPage.fillTestBoxForm(formData);
+        TextBoxFormPage.submitForm();
 
         $("#output").shouldHave(
                 text(formData.getFullName()),
